requirements.txt
Package              Version
-------------------- -------
asgiref              3.5.2
Django               4.1.1
django-debug-toolbar 3.6.0
django-environ       0.9.0
mysqlclient          2.1.1
pip                  22.2.2
setuptools           60.2.0
sqlparse             0.4.2
wheel                0.37.1

Django~=4.1.1
djangorestframework~=3.13.1